+++ xmake/actions/install/main.lua
@@ -73,7 +73,7 @@ function main()
                     cprint("${bright yellow}note: ${default yellow}try continue to install with administrator permission again?")
                     cprint("please input: y (y/n)")
 
-                    -- TODO read
+                    -- TODO fix read api 
                     -- get answer
                     io.flush()
                     if io._read() == 'y' then